+++ Runtime/Shaders/ShaderStrategyFactory.cs
@@ -40,6 +40,7 @@ namespace Bender_Dios.MenuRadial.Shaders
         private void InitializeStrategies()
         {
             RegisterStrategy(new LilToonShaderStrategy());
+            RegisterStrategy(new PoiyomiShaderStrategy());
         }
         
         /// <summary>